+++ common/src/main/java/me/pandamods/fallingtrees/exceptions/TreeException.java
@@ -1,0 +1,7 @@
+package me.pandamods.fallingtrees.exceptions;
+
+public abstract class TreeException extends Exception {
+	public TreeException(String message) {
+		super(message);
+	}
+}