+++ common/src/main/java/me/pandamods/fallingtrees/mixin/BlockBehaviourMixin.java
@@ -1,0 +1,19 @@
+package me.pandamods.fallingtrees.mixin;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import net.minecraft.world.level.block.state.BlockState;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+@Mixin(BlockBehaviour.class)
+public class BlockBehaviourMixin {
+	@Inject(method = "getDestroyProgress", at = @At("TAIL"), cancellable = true)
+	private void getDestroyProgress(BlockState state, Player player, BlockGetter level, BlockPos pos, CallbackInfoReturnable<Float> cir) {
+		// Todo: add dynamic breaking speed 
+	}
+}