+++ common/src/main/java/me/pandamods/fallingtrees/mixin/BlockStateMixin.java
@@ -29,13 +29,13 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(BlockBehaviour.BlockStateBase.class)
 public class BlockStateMixin {
-	@Inject(method = "getDestroyProgress", at = @At("RETURN"), cancellable = true)
-	public void getDestroyProgress(Player player, BlockGetter level, BlockPos pos, CallbackInfoReturnable<Float> cir) {
-		if (FallingTreesConfig.getCommonConfig().dynamicMiningSpeed.disable) return;
-		Tree<?> tree = TreeRegistry.getTree(level.getBlockState(pos));
-		if (tree != null && tree.willTreeFall(pos, level, player)) {
-			TreeData treeData = TreeCache.get(player, pos, () -> tree.getTreeData(new TreeDataBuilder(), pos, level));
-			cir.setReturnValue(cir.getReturnValueF() * treeData.miningSpeedMultiply());
-		}
-	}
+//	@Inject(method = "getDestroyProgress", at = @At("RETURN"), cancellable = true)
+//	public void getDestroyProgress(Player player, BlockGetter level, BlockPos pos, CallbackInfoReturnable<Float> cir) {
+//		if (FallingTreesConfig.getCommonConfig().dynamicMiningSpeed.disable) return;
+//		Tree<?> tree = TreeRegistry.getTree(level.getBlockState(pos));
+//		if (tree != null && tree.willTreeFall(pos, level, player)) {
+//			TreeData treeData = TreeCache.get(player, pos, () -> tree.getTreeData(new TreeDataBuilder(), pos, level));
+//			cir.setReturnValue(cir.getReturnValueF() * treeData.miningSpeedMultiply());
+//		}
+//	}
 }